1. Fixed side menu 2. added bottom sheet for user add 3. added vector icon library
