+++ App.js
@@ -1,10 +1,13 @@
 import {AuthProvider} from './src/contexts/AuthContext';
 import AppNavigation from './AppNavigation';
+import {GestureHandlerRootView} from 'react-native-gesture-handler';
 
 export default function App() {
   return (
-    <AuthProvider>
-      <AppNavigation />
-    </AuthProvider>
+    <GestureHandlerRootView style={{flex: 1}}>
+      <AuthProvider>
+        <AppNavigation />
+      </AuthProvider>
+    </GestureHandlerRootView>
   );
 }
