+++ package.json
@@ -10,6 +10,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@gorhom/bottom-sheet": "^4.5.1",
     "@react-native-async-storage/async-storage": "^1.21.0",
     "@react-native-community/masked-view": "^0.1.11",
     "@react-navigation/drawer": "^6.6.6",
@@ -26,6 +27,7 @@
     "react-native-safe-area-context": "^4.7.4",
     "react-native-screens": "^3.27.0",
     "react-native-svg": "^12.1.1",
+    "react-native-vector-icons": "^10.0.3",
     "react-navigation-drawer": "^2.7.2",
     "react-navigation-stack": "^2.10.4"
   },
