+++ src/screens/navigation/HomeStack.js
@@ -1,8 +1,10 @@
 import React from 'react';
-import {StyleSheet, Text, View} from 'react-native';
+import {Pressable, StyleSheet, Text, View} from 'react-native';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {DrawerScreenStack} from './DrawerStack';
 import {StatusBar} from 'native-base';
+import Ionicon from 'react-native-vector-icons/Ionicons';
+import {DrawerActions} from '@react-navigation/native';
 
 function DetailScreen() {
   return (
@@ -17,7 +19,19 @@ const HomeStack = createNativeStackNavigator();
 export function HomeScreenStack() {
   return (
     <HomeStack.Navigator>
-      <HomeStack.Screen name="DrawerHome" component={DrawerScreenStack} />
+      <HomeStack.Screen
+        name="DrawerHome"
+        component={DrawerScreenStack}
+        options={({navigation, route}) => ({
+          headerShown: false,
+          headerLeft: () => (
+            <Pressable
+              onPress={() => navigation.dispatch(DrawerActions.toggleDrawer())}>
+              <Ionicon name={'menu'} size={20} />
+            </Pressable>
+          ),
+        })}
+      />
       <HomeStack.Screen name="Detail" component={DetailScreen} />
     </HomeStack.Navigator>
   );
